move composite test utils
