+++ src/moonlink/src/storage/iceberg/state_tests.rs
@@ -37,6 +37,7 @@ use crate::storage::iceberg::test_utils::{
     load_arrow_batch, validate_recovered_snapshot,
 };
 use crate::storage::mooncake_table::delete_vector::BatchDeletionVector;
+use crate::storage::mooncake_table::state_test_utils::*;
 use crate::storage::mooncake_table::Snapshot;
 use crate::storage::mooncake_table::SnapshotOption;
 use crate::storage::MooncakeTable;
@@ -313,10 +314,7 @@ async fn test_state_1_3() -> IcebergResult<()> {
     table.append(row.clone()).unwrap();
 
     // Request to create snapshot.
-    table
-        .create_mooncake_and_iceberg_snapshot_for_test(&mut notify_rx)
-        .await
-        .unwrap();
+    create_mooncake_and_iceberg_snapshot_for_test(&mut table, &mut notify_rx).await;
 
     // Check iceberg snapshot status.
     let (next_file_id, snapshot) = iceberg_table_manager.load_snapshot_from_table().await?;
@@ -353,10 +351,7 @@ async fn test_state_1_4() -> IcebergResult<()> {
     table.delete(row.clone(), /*lsn=*/ 400).await;
 
     // Request to create snapshot.
-    table
-        .create_mooncake_and_iceberg_snapshot_for_test(&mut notify_rx)
-        .await
-        .unwrap();
+    create_mooncake_and_iceberg_snapshot_for_test(&mut table, &mut notify_rx).await;
 
     // Check iceberg snapshot status.
     let (next_file_id, snapshot) = iceberg_table_manager.load_snapshot_from_table().await?;
@@ -392,10 +387,7 @@ async fn test_state_1_5() -> IcebergResult<()> {
     table.append(row.clone()).unwrap();
 
     // Request to create snapshot.
-    table
-        .create_mooncake_and_iceberg_snapshot_for_test(&mut notify_rx)
-        .await
-        .unwrap();
+    create_mooncake_and_iceberg_snapshot_for_test(&mut table, &mut notify_rx).await;
 
     // Check iceberg snapshot status.
     let (next_file_id, snapshot) = iceberg_table_manager.load_snapshot_from_table().await?;
@@ -433,10 +425,7 @@ async fn test_state_1_6() -> IcebergResult<()> {
     table.delete(row.clone(), /*lsn=*/ 400).await;
 
     // Request to create snapshot.
-    table
-        .create_mooncake_and_iceberg_snapshot_for_test(&mut notify_rx)
-        .await
-        .unwrap();
+    create_mooncake_and_iceberg_snapshot_for_test(&mut table, &mut notify_rx).await;
 
     // Check iceberg snapshot status.
     let (next_file_id, snapshot) = iceberg_table_manager.load_snapshot_from_table().await?;
@@ -467,10 +456,7 @@ async fn test_state_2_1() -> IcebergResult<()> {
     table.commit(/*lsn=*/ 100);
 
     // Request to create snapshot.
-    table
-        .create_mooncake_and_iceberg_snapshot_for_test(&mut notify_rx)
-        .await
-        .unwrap();
+    create_mooncake_and_iceberg_snapshot_for_test(&mut table, &mut notify_rx).await;
 
     // Check iceberg snapshot status.
     let (next_file_id, snapshot) = iceberg_table_manager.load_snapshot_from_table().await?;
@@ -503,10 +489,7 @@ async fn test_state_2_2() -> IcebergResult<()> {
     table.delete(row.clone(), /*lsn=*/ 200).await;
 
     // Request to create snapshot.
-    table
-        .create_mooncake_and_iceberg_snapshot_for_test(&mut notify_rx)
-        .await
-        .unwrap();
+    create_mooncake_and_iceberg_snapshot_for_test(&mut table, &mut notify_rx).await;
 
     // Check iceberg snapshot status.
     let (next_file_id, snapshot) = iceberg_table_manager.load_snapshot_from_table().await?;
@@ -542,10 +525,7 @@ async fn test_state_2_3() -> IcebergResult<()> {
     table.commit(/*lsn=*/ 400);
 
     // Request to create snapshot.
-    table
-        .create_mooncake_and_iceberg_snapshot_for_test(&mut notify_rx)
-        .await
-        .unwrap();
+    create_mooncake_and_iceberg_snapshot_for_test(&mut table, &mut notify_rx).await;
 
     // Check iceberg snapshot status.
     let (next_file_id, snapshot) = iceberg_table_manager.load_snapshot_from_table().await?;
@@ -583,10 +563,7 @@ async fn test_state_2_4() -> IcebergResult<()> {
     table.delete(row.clone(), /*lsn=*/ 500).await;
 
     // Request to create snapshot.
-    table
-        .create_mooncake_and_iceberg_snapshot_for_test(&mut notify_rx)
-        .await
-        .unwrap();
+    create_mooncake_and_iceberg_snapshot_for_test(&mut table, &mut notify_rx).await;
 
     // Check iceberg snapshot status.
     let (next_file_id, snapshot) = iceberg_table_manager.load_snapshot_from_table().await?;
@@ -623,10 +600,7 @@ async fn test_state_2_5() -> IcebergResult<()> {
     table.commit(/*lsn=*/ 400);
 
     // Request to create snapshot.
-    table
-        .create_mooncake_and_iceberg_snapshot_for_test(&mut notify_rx)
-        .await
-        .unwrap();
+    create_mooncake_and_iceberg_snapshot_for_test(&mut table, &mut notify_rx).await;
 
     // Check iceberg snapshot status.
     let (next_file_id, snapshot) = iceberg_table_manager.load_snapshot_from_table().await?;
@@ -665,10 +639,7 @@ async fn test_state_2_6() -> IcebergResult<()> {
     table.delete(row.clone(), /*lsn=*/ 500).await;
 
     // Request to create snapshot.
-    table
-        .create_mooncake_and_iceberg_snapshot_for_test(&mut notify_rx)
-        .await
-        .unwrap();
+    create_mooncake_and_iceberg_snapshot_for_test(&mut table, &mut notify_rx).await;
 
     // Check iceberg snapshot status.
     let (next_file_id, snapshot) = iceberg_table_manager.load_snapshot_from_table().await?;
@@ -700,10 +671,7 @@ async fn test_state_3_1() -> IcebergResult<()> {
     table.flush(/*lsn=*/ 200).await.unwrap();
 
     // Request to create snapshot.
-    table
-        .create_mooncake_and_iceberg_snapshot_for_test(&mut notify_rx)
-        .await
-        .unwrap();
+    create_mooncake_and_iceberg_snapshot_for_test(&mut table, &mut notify_rx).await;
 
     // Check iceberg snapshot status.
     let (next_file_id, snapshot) = iceberg_table_manager.load_snapshot_from_table().await?;
@@ -737,10 +705,7 @@ async fn test_state_3_2() -> IcebergResult<()> {
     table.delete(row.clone(), /*lsn=*/ 300).await;
 
     // Request to create snapshot.
-    table
-        .create_mooncake_and_iceberg_snapshot_for_test(&mut notify_rx)
-        .await
-        .unwrap();
+    create_mooncake_and_iceberg_snapshot_for_test(&mut table, &mut notify_rx).await;
 
     // Check iceberg snapshot status.
     let (next_file_id, snapshot) = iceberg_table_manager.load_snapshot_from_table().await?;
@@ -777,10 +742,7 @@ async fn test_state_3_3_deletion_before_flush() -> IcebergResult<()> {
     table.flush(/*lsn=*/ 500).await.unwrap();
 
     // Request to create snapshot.
-    table
-        .create_mooncake_and_iceberg_snapshot_for_test(&mut notify_rx)
-        .await
-        .unwrap();
+    create_mooncake_and_iceberg_snapshot_for_test(&mut table, &mut notify_rx).await;
 
     // Check iceberg snapshot status.
     let (next_file_id, snapshot) = iceberg_table_manager.load_snapshot_from_table().await?;
@@ -822,10 +784,7 @@ async fn test_state_3_3_deletion_after_flush() -> IcebergResult<()> {
     table.commit(/*lsn=*/ 500);
 
     // Request to create snapshot.
-    table
-        .create_mooncake_and_iceberg_snapshot_for_test(&mut notify_rx)
-        .await
-        .unwrap();
+    create_mooncake_and_iceberg_snapshot_for_test(&mut table, &mut notify_rx).await;
 
     // Check iceberg snapshot status.
     let (next_file_id, snapshot) = iceberg_table_manager.load_snapshot_from_table().await?;
@@ -869,10 +828,7 @@ async fn test_state_3_4_committed_deletion_before_flush() -> IcebergResult<()> {
     table.delete(row.clone(), /*lsn=*/ 600).await;
 
     // Request to create snapshot.
-    table
-        .create_mooncake_and_iceberg_snapshot_for_test(&mut notify_rx)
-        .await
-        .unwrap();
+    create_mooncake_and_iceberg_snapshot_for_test(&mut table, &mut notify_rx).await;
 
     // Check iceberg snapshot status.
     let (next_file_id, snapshot) = iceberg_table_manager.load_snapshot_from_table().await?;
@@ -916,10 +872,7 @@ async fn test_state_3_4_committed_deletion_after_flush() -> IcebergResult<()> {
     table.delete(row.clone(), /*lsn=*/ 600).await;
 
     // Request to create snapshot.
-    table
-        .create_mooncake_and_iceberg_snapshot_for_test(&mut notify_rx)
-        .await
-        .unwrap();
+    create_mooncake_and_iceberg_snapshot_for_test(&mut table, &mut notify_rx).await;
 
     // Check iceberg snapshot status.
     let (next_file_id, snapshot) = iceberg_table_manager.load_snapshot_from_table().await?;
@@ -962,10 +915,7 @@ async fn test_state_3_5() -> IcebergResult<()> {
     table.flush(/*lsn=*/ 400).await.unwrap();
 
     // Request to create snapshot.
-    table
-        .create_mooncake_and_iceberg_snapshot_for_test(&mut notify_rx)
-        .await
-        .unwrap();
+    create_mooncake_and_iceberg_snapshot_for_test(&mut table, &mut notify_rx).await;
 
     // Check iceberg snapshot status.
     let (next_file_id, snapshot) = iceberg_table_manager.load_snapshot_from_table().await?;
@@ -1010,10 +960,7 @@ async fn test_state_3_6() -> IcebergResult<()> {
     table.delete(row.clone(), /*lsn=*/ 500).await;
 
     // Request to create snapshot.
-    table
-        .create_mooncake_and_iceberg_snapshot_for_test(&mut notify_rx)
-        .await
-        .unwrap();
+    create_mooncake_and_iceberg_snapshot_for_test(&mut table, &mut notify_rx).await;
 
     // Check iceberg snapshot status.
     let (next_file_id, snapshot) = iceberg_table_manager.load_snapshot_from_table().await?;
@@ -1056,10 +1003,7 @@ async fn test_state_4_1() -> IcebergResult<()> {
     table.append(row).unwrap();
 
     // Request to create snapshot.
-    table
-        .create_mooncake_and_iceberg_snapshot_for_test(&mut notify_rx)
-        .await
-        .unwrap();
+    create_mooncake_and_iceberg_snapshot_for_test(&mut table, &mut notify_rx).await;
 
     // Check iceberg snapshot status.
     let (next_file_id, snapshot) = iceberg_table_manager.load_snapshot_from_table().await?;
@@ -1099,10 +1043,7 @@ async fn test_state_4_2() -> IcebergResult<()> {
     table.append(row).unwrap();
 
     // Request to create snapshot.
-    table
-        .create_mooncake_and_iceberg_snapshot_for_test(&mut notify_rx)
-        .await
-        .unwrap();
+    create_mooncake_and_iceberg_snapshot_for_test(&mut table, &mut notify_rx).await;
 
     // Check iceberg snapshot status.
     let (next_file_id, snapshot) = iceberg_table_manager.load_snapshot_from_table().await?;
@@ -1145,10 +1086,7 @@ async fn test_state_4_3() -> IcebergResult<()> {
     table.append(row).unwrap();
 
     // Request to create snapshot.
-    table
-        .create_mooncake_and_iceberg_snapshot_for_test(&mut notify_rx)
-        .await
-        .unwrap();
+    create_mooncake_and_iceberg_snapshot_for_test(&mut table, &mut notify_rx).await;
 
     // Check iceberg snapshot status.
     let (next_file_id, snapshot) = iceberg_table_manager.load_snapshot_from_table().await?;
@@ -1193,10 +1131,7 @@ async fn test_state_4_4() -> IcebergResult<()> {
     table.append(row).unwrap();
 
     // Request to create snapshot.
-    table
-        .create_mooncake_and_iceberg_snapshot_for_test(&mut notify_rx)
-        .await
-        .unwrap();
+    create_mooncake_and_iceberg_snapshot_for_test(&mut table, &mut notify_rx).await;
 
     // Check iceberg snapshot status.
     let (next_file_id, snapshot) = iceberg_table_manager.load_snapshot_from_table().await?;
@@ -1240,10 +1175,7 @@ async fn test_state_4_5() -> IcebergResult<()> {
     table.append(row).unwrap();
 
     // Request to create snapshot.
-    table
-        .create_mooncake_and_iceberg_snapshot_for_test(&mut notify_rx)
-        .await
-        .unwrap();
+    create_mooncake_and_iceberg_snapshot_for_test(&mut table, &mut notify_rx).await;
 
     // Check iceberg snapshot status.
     let (next_file_id, snapshot) = iceberg_table_manager.load_snapshot_from_table().await?;
@@ -1289,10 +1221,7 @@ async fn test_state_4_6() -> IcebergResult<()> {
     table.delete(row.clone(), /*lsn=*/ 500).await;
 
     // Request to create snapshot.
-    table
-        .create_mooncake_and_iceberg_snapshot_for_test(&mut notify_rx)
-        .await
-        .unwrap();
+    create_mooncake_and_iceberg_snapshot_for_test(&mut table, &mut notify_rx).await;
 
     // Check iceberg snapshot status.
     let (next_file_id, snapshot) = iceberg_table_manager.load_snapshot_from_table().await?;
@@ -1332,10 +1261,7 @@ async fn test_state_5_1() -> IcebergResult<()> {
     table.commit(/*lsn=*/ 300);
 
     // Request to create snapshot.
-    table
-        .create_mooncake_and_iceberg_snapshot_for_test(&mut notify_rx)
-        .await
-        .unwrap();
+    create_mooncake_and_iceberg_snapshot_for_test(&mut table, &mut notify_rx).await;
 
     // Check iceberg snapshot status.
     let (next_file_id, snapshot) = iceberg_table_manager.load_snapshot_from_table().await?;
@@ -1377,10 +1303,7 @@ async fn test_state_5_2() -> IcebergResult<()> {
     table.commit(/*lsn=*/ 400);
 
     // Request to create snapshot.
-    table
-        .create_mooncake_and_iceberg_snapshot_for_test(&mut notify_rx)
-        .await
-        .unwrap();
+    create_mooncake_and_iceberg_snapshot_for_test(&mut table, &mut notify_rx).await;
 
     // Check iceberg snapshot status.
     let (next_file_id, snapshot) = iceberg_table_manager.load_snapshot_from_table().await?;
@@ -1425,10 +1348,7 @@ async fn test_state_5_3_deletion_before_flush() -> IcebergResult<()> {
     table.commit(/*lsn=*/ 600);
 
     // Request to create snapshot.
-    table
-        .create_mooncake_and_iceberg_snapshot_for_test(&mut notify_rx)
-        .await
-        .unwrap();
+    create_mooncake_and_iceberg_snapshot_for_test(&mut table, &mut notify_rx).await;
 
     // Check iceberg snapshot status.
     let (next_file_id, snapshot) = iceberg_table_manager.load_snapshot_from_table().await?;
@@ -1478,10 +1398,7 @@ async fn test_state_5_3_deletion_after_flush() -> IcebergResult<()> {
     table.commit(/*lsn=*/ 600);
 
     // Request to create snapshot.
-    table
-        .create_mooncake_and_iceberg_snapshot_for_test(&mut notify_rx)
-        .await
-        .unwrap();
+    create_mooncake_and_iceberg_snapshot_for_test(&mut table, &mut notify_rx).await;
 
     // Check iceberg snapshot status.
     let (next_file_id, snapshot) = iceberg_table_manager.load_snapshot_from_table().await?;
@@ -1533,10 +1450,7 @@ async fn test_state_5_4_committed_deletion_before_flush() -> IcebergResult<()> {
     table.commit(/*lsn=*/ 700);
 
     // Request to create snapshot.
-    table
-        .create_mooncake_and_iceberg_snapshot_for_test(&mut notify_rx)
-        .await
-        .unwrap();
+    create_mooncake_and_iceberg_snapshot_for_test(&mut table, &mut notify_rx).await;
 
     // Check iceberg snapshot status.
     let (next_file_id, snapshot) = iceberg_table_manager.load_snapshot_from_table().await?;
@@ -1588,10 +1502,7 @@ async fn test_state_5_4_committed_deletion_after_flush() -> IcebergResult<()> {
     table.commit(/*lsn=*/ 700);
 
     // Request to create snapshot.
-    table
-        .create_mooncake_and_iceberg_snapshot_for_test(&mut notify_rx)
-        .await
-        .unwrap();
+    create_mooncake_and_iceberg_snapshot_for_test(&mut table, &mut notify_rx).await;
 
     // Check iceberg snapshot status.
     let (next_file_id, snapshot) = iceberg_table_manager.load_snapshot_from_table().await?;
@@ -1642,10 +1553,7 @@ async fn test_state_5_5() -> IcebergResult<()> {
     table.commit(/*lsn=*/ 500);
 
     // Request to create snapshot.
-    table
-        .create_mooncake_and_iceberg_snapshot_for_test(&mut notify_rx)
-        .await
-        .unwrap();
+    create_mooncake_and_iceberg_snapshot_for_test(&mut table, &mut notify_rx).await;
 
     // Check iceberg snapshot status.
     let (next_file_id, snapshot) = iceberg_table_manager.load_snapshot_from_table().await?;
@@ -1698,10 +1606,7 @@ async fn test_state_5_6() -> IcebergResult<()> {
     table.delete(row.clone(), /*lsn=*/ 600).await;
 
     // Request to create snapshot.
-    table
-        .create_mooncake_and_iceberg_snapshot_for_test(&mut notify_rx)
-        .await
-        .unwrap();
+    create_mooncake_and_iceberg_snapshot_for_test(&mut table, &mut notify_rx).await;
 
     // Check iceberg snapshot status.
     let (next_file_id, snapshot) = iceberg_table_manager.load_snapshot_from_table().await?;
@@ -1753,10 +1658,7 @@ async fn test_state_6_1() -> IcebergResult<()> {
     table.append(row).unwrap();
 
     // Request to create snapshot.
-    table
-        .create_mooncake_and_iceberg_snapshot_for_test(&mut notify_rx)
-        .await
-        .unwrap();
+    create_mooncake_and_iceberg_snapshot_for_test(&mut table, &mut notify_rx).await;
 
     // Check iceberg snapshot status.
     let (next_file_id, snapshot) = iceberg_table_manager.load_snapshot_from_table().await?;
@@ -1805,10 +1707,7 @@ async fn test_state_6_2() -> IcebergResult<()> {
     table.append(row).unwrap();
 
     // Request to create snapshot.
-    table
-        .create_mooncake_and_iceberg_snapshot_for_test(&mut notify_rx)
-        .await
-        .unwrap();
+    create_mooncake_and_iceberg_snapshot_for_test(&mut table, &mut notify_rx).await;
 
     // Check iceberg snapshot status.
     let (next_file_id, snapshot) = iceberg_table_manager.load_snapshot_from_table().await?;
@@ -1860,10 +1759,7 @@ async fn test_state_6_3_deletion_before_flush() -> IcebergResult<()> {
     table.append(row).unwrap();
 
     // Request to create snapshot.
-    table
-        .create_mooncake_and_iceberg_snapshot_for_test(&mut notify_rx)
-        .await
-        .unwrap();
+    create_mooncake_and_iceberg_snapshot_for_test(&mut table, &mut notify_rx).await;
 
     // Check iceberg snapshot status.
     let (next_file_id, snapshot) = iceberg_table_manager.load_snapshot_from_table().await?;
@@ -1920,10 +1816,7 @@ async fn test_state_6_3_deletion_after_flush() -> IcebergResult<()> {
     table.append(row).unwrap();
 
     // Request to create snapshot.
-    table
-        .create_mooncake_and_iceberg_snapshot_for_test(&mut notify_rx)
-        .await
-        .unwrap();
+    create_mooncake_and_iceberg_snapshot_for_test(&mut table, &mut notify_rx).await;
 
     // Check iceberg snapshot status.
     let (next_file_id, snapshot) = iceberg_table_manager.load_snapshot_from_table().await?;
@@ -1981,10 +1874,7 @@ async fn test_state_6_4_committed_deletion_before_flush() -> IcebergResult<()> {
     table.append(row).unwrap();
 
     // Request to create snapshot.
-    table
-        .create_mooncake_and_iceberg_snapshot_for_test(&mut notify_rx)
-        .await
-        .unwrap();
+    create_mooncake_and_iceberg_snapshot_for_test(&mut table, &mut notify_rx).await;
 
     // Check iceberg snapshot status.
     let (next_file_id, snapshot) = iceberg_table_manager.load_snapshot_from_table().await?;
@@ -2043,10 +1933,7 @@ async fn test_state_6_4_committed_deletion_after_flush() -> IcebergResult<()> {
     table.append(row).unwrap();
 
     // Request to create snapshot.
-    table
-        .create_mooncake_and_iceberg_snapshot_for_test(&mut notify_rx)
-        .await
-        .unwrap();
+    create_mooncake_and_iceberg_snapshot_for_test(&mut table, &mut notify_rx).await;
 
     // Check iceberg snapshot status.
     let (next_file_id, snapshot) = iceberg_table_manager.load_snapshot_from_table().await?;
@@ -2104,10 +1991,7 @@ async fn test_state_6_5() -> IcebergResult<()> {
     table.append(row).unwrap();
 
     // Request to create snapshot.
-    table
-        .create_mooncake_and_iceberg_snapshot_for_test(&mut notify_rx)
-        .await
-        .unwrap();
+    create_mooncake_and_iceberg_snapshot_for_test(&mut table, &mut notify_rx).await;
 
     // Check iceberg snapshot status.
     let (next_file_id, snapshot) = iceberg_table_manager.load_snapshot_from_table().await?;
@@ -2167,10 +2051,7 @@ async fn test_state_6_6() -> IcebergResult<()> {
     table.delete(/*row=*/ row.clone(), /*lsn=*/ 600).await;
 
     // Request to create snapshot.
-    table
-        .create_mooncake_and_iceberg_snapshot_for_test(&mut notify_rx)
-        .await
-        .unwrap();
+    create_mooncake_and_iceberg_snapshot_for_test(&mut table, &mut notify_rx).await;
 
     // Check iceberg snapshot status.
     let (next_file_id, snapshot) = iceberg_table_manager.load_snapshot_from_table().await?;
@@ -2197,10 +2078,7 @@ async fn test_state_7_1() -> IcebergResult<()> {
         create_table_and_iceberg_manager(&temp_dir).await;
 
     // Request to create snapshot.
-    table
-        .create_mooncake_and_iceberg_snapshot_for_test(&mut notify_rx)
-        .await
-        .unwrap();
+    create_mooncake_and_iceberg_snapshot_for_test(&mut table, &mut notify_rx).await;
 
     // Check iceberg snapshot status.
     let (next_file_id, snapshot) = iceberg_table_manager.load_snapshot_from_table().await?;
@@ -2227,10 +2105,7 @@ async fn test_state_7_2() -> IcebergResult<()> {
     table.delete(/*row=*/ old_row.clone(), /*lsn=*/ 200).await;
 
     // Request to create snapshot.
-    table
-        .create_mooncake_and_iceberg_snapshot_for_test(&mut notify_rx)
-        .await
-        .unwrap();
+    create_mooncake_and_iceberg_snapshot_for_test(&mut table, &mut notify_rx).await;
 
     // Check iceberg snapshot status.
     let (next_file_id, snapshot) = iceberg_table_manager.load_snapshot_from_table().await?;
@@ -2258,10 +2133,7 @@ async fn test_state_7_3() -> IcebergResult<()> {
     table.commit(/*lsn=*/ 300);
 
     // Request to create snapshot.
-    table
-        .create_mooncake_and_iceberg_snapshot_for_test(&mut notify_rx)
-        .await
-        .unwrap();
+    create_mooncake_and_iceberg_snapshot_for_test(&mut table, &mut notify_rx).await;
 
     // Check iceberg snapshot status.
     let (next_file_id, snapshot) = iceberg_table_manager.load_snapshot_from_table().await?;
@@ -2292,10 +2164,7 @@ async fn test_state_7_4() -> IcebergResult<()> {
     table.delete(old_row_2, /*lsn=*/ 400).await;
 
     // Request to create snapshot.
-    table
-        .create_mooncake_and_iceberg_snapshot_for_test(&mut notify_rx)
-        .await
-        .unwrap();
+    create_mooncake_and_iceberg_snapshot_for_test(&mut table, &mut notify_rx).await;
 
     // Check iceberg snapshot status.
     let (next_file_id, snapshot) = iceberg_table_manager.load_snapshot_from_table().await?;
@@ -2324,10 +2193,7 @@ async fn test_state_7_5() -> IcebergResult<()> {
     table.flush(/*lsn=*/ 300).await.unwrap();
 
     // Request to create snapshot.
-    table
-        .create_mooncake_and_iceberg_snapshot_for_test(&mut notify_rx)
-        .await
-        .unwrap();
+    create_mooncake_and_iceberg_snapshot_for_test(&mut table, &mut notify_rx).await;
 
     // Check iceberg snapshot status.
     let (next_file_id, snapshot) = iceberg_table_manager.load_snapshot_from_table().await?;
@@ -2358,10 +2224,7 @@ async fn test_state_7_6() -> IcebergResult<()> {
     table.delete(old_row_2.clone(), /*lsn=*/ 400).await;
 
     // Request to create snapshot.
-    table
-        .create_mooncake_and_iceberg_snapshot_for_test(&mut notify_rx)
-        .await
-        .unwrap();
+    create_mooncake_and_iceberg_snapshot_for_test(&mut table, &mut notify_rx).await;
 
     // Check iceberg snapshot status.
     let (next_file_id, snapshot) = iceberg_table_manager.load_snapshot_from_table().await?;
