+++ src/moonlink/src/storage/mooncake_table.rs
@@ -38,6 +38,12 @@ pub(crate) use crate::storage::mooncake_table::table_snapshot::{
 use crate::storage::storage_utils::{FileId, TableId};
 use crate::table_notify::TableNotify;
 use crate::NonEvictableHandle;
+use arrow::record_batch::RecordBatch;
+use arrow_schema::Schema;
+use delete_vector::BatchDeletionVector;
+pub(crate) use disk_slice::DiskSliceWriter;
+use mem_slice::MemSlice;
+pub(crate) use snapshot::{PuffinDeletionBlobAtRead, SnapshotTableState};
 use std::collections::HashMap;
 use std::path::PathBuf;
 use std::sync::Arc;
@@ -45,19 +51,10 @@ use table_snapshot::{IcebergSnapshotImportResult, IcebergSnapshotIndexMergeResul
 #[cfg(test)]
 use tokio::sync::mpsc::Receiver;
 use tokio::sync::mpsc::Sender;
+use tokio::sync::{watch, RwLock};
 use tracing::info_span;
 use tracing::Instrument;
 use transaction_stream::{TransactionStreamOutput, TransactionStreamState};
-
-#[cfg(test)]
-use crate::storage::io_utils;
-use arrow::record_batch::RecordBatch;
-use arrow_schema::Schema;
-use delete_vector::BatchDeletionVector;
-pub(crate) use disk_slice::DiskSliceWriter;
-use mem_slice::MemSlice;
-pub(crate) use snapshot::{PuffinDeletionBlobAtRead, SnapshotTableState};
-use tokio::sync::{watch, RwLock};
 
 // TODO(hjiang): Add another threshold for merged file indices to trigger iceberg snapshot.
 #[derive(Clone, Debug)]
@@ -974,284 +971,6 @@ impl MooncakeTable {
             })
             .await
             .unwrap();
-    }
-
-    // ================================
-    // Test util functions
-    // ================================
-    //
-    // Test util function to block wait and get iceberg / file indices merge payload.
-    #[cfg(test)]
-    async fn sync_mooncake_snapshot(
-        receiver: &mut Receiver<TableNotify>,
-    ) -> (
-        Option<IcebergSnapshotPayload>,
-        Option<FileIndiceMergePayload>,
-        Option<DataCompactionPayload>,
-        Vec<String>,
-    ) {
-        let notification = receiver.recv().await.unwrap();
-        if let TableNotify::MooncakeTableSnapshot {
-            iceberg_snapshot_payload,
-            file_indice_merge_payload,
-            data_compaction_payload,
-            evicted_data_files_to_delete,
-            ..
-        } = notification
-        {
-            (
-                iceberg_snapshot_payload,
-                file_indice_merge_payload,
-                data_compaction_payload,
-                evicted_data_files_to_delete,
-            )
-        } else {
-            panic!("Expected mooncake snapshot completion notification, but get others.");
-        }
-    }
-
-    #[cfg(test)]
-    async fn sync_iceberg_snapshot(
-        receiver: &mut Receiver<TableNotify>,
-    ) -> Result<IcebergSnapshotResult> {
-        let notification = receiver.recv().await.unwrap();
-        if let TableNotify::IcebergSnapshot {
-            iceberg_snapshot_result,
-        } = notification
-        {
-            iceberg_snapshot_result
-        } else {
-            panic!("Expected iceberg completion snapshot notification, but get mooncake one.");
-        }
-    }
-
-    #[cfg(test)]
-    async fn sync_index_merge(receiver: &mut Receiver<TableNotify>) -> FileIndiceMergeResult {
-        let notification = receiver.recv().await.unwrap();
-        if let TableNotify::IndexMerge { index_merge_result } = notification {
-            index_merge_result
-        } else {
-            panic!("Expected index merge completion notification, but get another one.");
-        }
-    }
-
-    #[cfg(test)]
-    async fn sync_data_compaction(
-        receiver: &mut Receiver<TableNotify>,
-    ) -> Result<DataCompactionResult> {
-        let notification = receiver.recv().await.unwrap();
-        if let TableNotify::DataCompaction {
-            data_compaction_result,
-        } = notification
-        {
-            data_compaction_result
-        } else {
-            panic!("Expected data compaction completion notification, but get mooncake one.");
-        }
-    }
-
-    // Test util function, which creates mooncake snapshot for testing.
-    #[cfg(test)]
-    pub(crate) async fn create_mooncake_snapshot_for_test(
-        &mut self,
-        receiver: &mut Receiver<TableNotify>,
-    ) -> (
-        Option<IcebergSnapshotPayload>,
-        Option<FileIndiceMergePayload>,
-        Option<DataCompactionPayload>,
-        Vec<String>,
-    ) {
-        let mooncake_snapshot_created = self.create_snapshot(SnapshotOption {
-            force_create: true,
-            skip_iceberg_snapshot: false,
-            skip_data_file_compaction: false,
-            skip_file_indices_merge: false,
-        });
-        assert!(mooncake_snapshot_created);
-        Self::sync_mooncake_snapshot(receiver).await
-    }
-
-    // Test util function, which updates mooncake table snapshot and create iceberg snapshot in a serial fashion.
-    #[cfg(test)]
-    pub(crate) async fn create_mooncake_and_iceberg_snapshot_for_test(
-        &mut self,
-        receiver: &mut Receiver<TableNotify>,
-    ) -> Result<()> {
-        // Create mooncake snapshot and block wait completion.
-        let (iceberg_snapshot_payload, _, _, evicted_data_files_to_delete) =
-            self.create_mooncake_snapshot_for_test(receiver).await;
-
-        // Delete evicted object storage cache entries immediately to make sure later accesses all happen on persisted files.
-        io_utils::delete_local_files(evicted_data_files_to_delete)
-            .await
-            .unwrap();
-
-        // Create iceberg snapshot if possible.
-        if let Some(iceberg_snapshot_payload) = iceberg_snapshot_payload {
-            self.persist_iceberg_snapshot(iceberg_snapshot_payload);
-            let iceberg_snapshot_result = Self::sync_iceberg_snapshot(receiver).await;
-            self.set_iceberg_snapshot_res(iceberg_snapshot_result.unwrap());
-        }
-
-        Ok(())
-    }
-
-    // Test util to block wait current mooncake snapshot completion, get the iceberg persistence payload, and perform a new mooncake snapshot and wait completion.
-    #[cfg(test)]
-    async fn sync_mooncake_snapshot_and_create_new_by_iceberg_payload(
-        &mut self,
-        receiver: &mut Receiver<TableNotify>,
-    ) {
-        let (iceberg_snapshot_payload, _, _, evicted_data_files_to_delete) =
-            Self::sync_mooncake_snapshot(receiver).await;
-        // Delete evicted object storage cache entries immediately to make sure later accesses all happen on persisted files.
-        io_utils::delete_local_files(evicted_data_files_to_delete)
-            .await
-            .unwrap();
-
-        let iceberg_snapshot_payload = iceberg_snapshot_payload.unwrap();
-        self.persist_iceberg_snapshot(iceberg_snapshot_payload);
-        let iceberg_snapshot_result = Self::sync_iceberg_snapshot(receiver).await;
-        self.set_iceberg_snapshot_res(iceberg_snapshot_result.unwrap());
-
-        // Create mooncake snapshot after buffering iceberg snapshot result, to make sure mooncake snapshot is at a consistent state.
-        assert!(self.create_snapshot(SnapshotOption {
-            force_create: true,
-            skip_iceberg_snapshot: true,
-            skip_file_indices_merge: true,
-            skip_data_file_compaction: true,
-        }));
-    }
-
-    // Test util function, which does the following things in serial fashion.
-    // (1) updates mooncake table snapshot, (2) create iceberg snapshot, (3) trigger data compaction, (4) perform data compaction, (5) another mooncake and iceberg snapshot.
-    //
-    // # Arguments
-    //
-    // * injected_committed_deletion_rows: rows to delete and commit in between data compaction initiation and snapshot creation
-    // * injected_uncommitted_deletion_rows: rows to delete but not commit in between data compaction initiation and snapshot creation
-    #[cfg(test)]
-    pub(crate) async fn create_mooncake_and_iceberg_snapshot_for_data_compaction_for_test(
-        &mut self,
-        receiver: &mut Receiver<TableNotify>,
-        injected_committed_deletion_rows: Vec<(MoonlinkRow, u64 /*lsn*/)>,
-        injected_uncommitted_deletion_rows: Vec<(MoonlinkRow, u64 /*lsn*/)>,
-    ) -> Result<()> {
-        // Create mooncake snapshot.
-        let force_snapshot_option = SnapshotOption {
-            force_create: true,
-            skip_iceberg_snapshot: false,
-            skip_file_indices_merge: true,
-            skip_data_file_compaction: false,
-        };
-        assert!(self.create_snapshot(force_snapshot_option.clone()));
-
-        // Create iceberg snapshot.
-        let (iceberg_snapshot_payload, _, _, evicted_data_files_to_delete) =
-            Self::sync_mooncake_snapshot(receiver).await;
-        // Delete evicted object storage cache entries immediately to make sure later accesses all happen on persisted files.
-        io_utils::delete_local_files(evicted_data_files_to_delete)
-            .await
-            .unwrap();
-
-        if let Some(iceberg_snapshot_payload) = iceberg_snapshot_payload {
-            self.persist_iceberg_snapshot(iceberg_snapshot_payload);
-            let iceberg_snapshot_result = Self::sync_iceberg_snapshot(receiver).await;
-            self.set_iceberg_snapshot_res(iceberg_snapshot_result.unwrap());
-        }
-
-        // Get data compaction payload.
-        assert!(self.create_snapshot(force_snapshot_option.clone()));
-        let (iceberg_snapshot_payload, _, data_compaction_payload, evicted_data_files_to_delete) =
-            Self::sync_mooncake_snapshot(receiver).await;
-        // Delete evicted object storage cache entries immediately to make sure later accesses all happen on persisted files.
-        io_utils::delete_local_files(evicted_data_files_to_delete)
-            .await
-            .unwrap();
-
-        assert!(iceberg_snapshot_payload.is_none());
-        let data_compaction_payload = data_compaction_payload.unwrap();
-
-        // Perform and block wait data compaction.
-        self.perform_data_compaction(data_compaction_payload);
-        let data_compaction_result = Self::sync_data_compaction(receiver).await.unwrap();
-
-        // Before create snapshot for compaction results, perform another deletion operations.
-        for (cur_row, lsn) in injected_committed_deletion_rows {
-            self.delete(cur_row, lsn).await;
-            self.commit(/*lsn=*/ lsn);
-        }
-        for (cur_row, lsn) in injected_uncommitted_deletion_rows {
-            self.delete(cur_row, lsn).await;
-        }
-
-        // Set data compaction result and trigger another iceberg snapshot.
-        self.set_data_compaction_res(data_compaction_result);
-        assert!(self.create_snapshot(SnapshotOption {
-            force_create: true,
-            skip_iceberg_snapshot: false,
-            skip_file_indices_merge: true,
-            skip_data_file_compaction: false,
-        }));
-        self.sync_mooncake_snapshot_and_create_new_by_iceberg_payload(receiver).await;
-
-        Ok(())
-    }
-
-    // Test util function, which does the following things in serial fashion.
-    // (1) updates mooncake table snapshot, (2) create iceberg snapshot, (3) trigger index merge, (4) perform index merge, (5) another mooncake and iceberg snapshot.
-    #[cfg(test)]
-    pub(crate) async fn create_mooncake_and_iceberg_snapshot_for_index_merge_for_test(
-        &mut self,
-        receiver: &mut Receiver<TableNotify>,
-    ) -> Result<()> {
-        // Create mooncake snapshot.
-        let force_snapshot_option = SnapshotOption {
-            force_create: true,
-            skip_iceberg_snapshot: false,
-            skip_file_indices_merge: false,
-            skip_data_file_compaction: false,
-        };
-        assert!(self.create_snapshot(force_snapshot_option.clone()));
-
-        // Create iceberg snapshot.
-        let (iceberg_snapshot_payload, _, _, evicted_data_files_to_delete) =
-            Self::sync_mooncake_snapshot(receiver).await;
-        // Delete evicted object storage cache entries immediately to make sure later accesses all happen on persisted files.
-        io_utils::delete_local_files(evicted_data_files_to_delete)
-            .await
-            .unwrap();
-
-        if let Some(iceberg_snapshot_payload) = iceberg_snapshot_payload {
-            self.persist_iceberg_snapshot(iceberg_snapshot_payload);
-            let iceberg_snapshot_result = Self::sync_iceberg_snapshot(receiver).await;
-            self.set_iceberg_snapshot_res(iceberg_snapshot_result.unwrap());
-        }
-
-        // Perform index merge.
-        assert!(self.create_snapshot(force_snapshot_option.clone()));
-        let (iceberg_snapshot_payload, file_indice_merge_payload, _, evicted_data_files_to_delete) =
-            Self::sync_mooncake_snapshot(receiver).await;
-        // Delete evicted object storage cache entries immediately to make sure later accesses all happen on persisted files.
-        io_utils::delete_local_files(evicted_data_files_to_delete)
-            .await
-            .unwrap();
-
-        assert!(iceberg_snapshot_payload.is_none());
-        let file_indice_merge_payload = file_indice_merge_payload.unwrap();
-
-        self.perform_index_merge(file_indice_merge_payload);
-        let index_merge_result = Self::sync_index_merge(receiver).await;
-        self.set_file_indices_merge_res(index_merge_result);
-        assert!(self.create_snapshot(SnapshotOption {
-            force_create: true,
-            skip_iceberg_snapshot: false,
-            skip_file_indices_merge: false,
-            skip_data_file_compaction: false,
-        }));
-        self.sync_mooncake_snapshot_and_create_new_by_iceberg_payload(receiver).await;
-
-        Ok(())
     }
 }
