+++ src/moonlink/src/storage/mooncake_table/file_index_state_tests.rs
@@ -89,13 +89,9 @@ async fn test_1_recover_3() {
 
     let (mut table, mut table_notify) =
         prepare_test_disk_file(&temp_dir, ObjectStorageCache::new(cache_config)).await;
-    table
-        .create_mooncake_and_iceberg_snapshot_for_test(&mut table_notify)
-        .await
-        .unwrap();
-    let (_, _, _, files_to_delete) = table
-        .create_mooncake_snapshot_for_test(&mut table_notify)
-        .await;
+    create_mooncake_and_iceberg_snapshot_for_test(&mut table, &mut table_notify).await;
+    let (_, _, _, files_to_delete) =
+        create_mooncake_snapshot_for_test(&mut table, &mut table_notify).await;
     assert!(files_to_delete.is_empty());
 
     // Now the disk file and deletion vector has been persist into iceberg.
@@ -265,13 +261,9 @@ async fn test_3_index_merge() {
 
     let (mut table, mut table_notify) =
         prepare_test_disk_files_for_index_merge(&temp_dir, object_storage_cache.clone()).await;
-    table
-        .create_mooncake_and_iceberg_snapshot_for_test(&mut table_notify)
-        .await
-        .unwrap();
-    let (_, index_merge_payload, _, files_to_delete) = table
-        .create_mooncake_snapshot_for_test(&mut table_notify)
-        .await;
+    create_mooncake_and_iceberg_snapshot_for_test(&mut table, &mut table_notify).await;
+    let (_, index_merge_payload, _, files_to_delete) =
+        create_mooncake_snapshot_for_test(&mut table, &mut table_notify).await;
     assert!(files_to_delete.is_empty());
     let index_merge_payload = index_merge_payload.unwrap();
