+++ src/moonlink/src/storage/mooncake_table/state_test_utils.rs
@@ -5,6 +5,8 @@ use tempfile::TempDir;
 use tokio::sync::mpsc;
 use tokio::sync::mpsc::Receiver;
 
+#[cfg(test)]
+use crate::row::MoonlinkRow;
 use crate::storage::cache::object_storage::base_cache::CacheTrait;
 use crate::storage::cache::object_storage::base_cache::{CacheEntry, FileMetadata};
 use crate::storage::compaction::compaction_config::DataCompactionConfig;
@@ -12,7 +14,8 @@ use crate::storage::iceberg::test_utils::*;
 use crate::storage::index::persisted_bucket_hash_map::GlobalIndex;
 use crate::storage::io_utils;
 use crate::storage::mooncake_table::{
-    DataCompactionPayload, DiskFileEntry, FileIndiceMergePayload, SnapshotOption,
+    DataCompactionPayload, DataCompactionResult, DiskFileEntry, FileIndiceMergePayload,
+    FileIndiceMergeResult, IcebergSnapshotPayload, IcebergSnapshotResult, SnapshotOption,
     TableConfig as MooncakeTableConfig,
 };
 use crate::storage::storage_utils::{
@@ -91,7 +94,7 @@ pub(super) async fn drop_read_states_and_create_mooncake_snapshot(
     receiver: &mut Receiver<TableNotify>,
 ) -> Vec<String> {
     drop_read_states(read_states, table, receiver).await;
-    let (_, _, _, files_to_delete) = table.create_mooncake_snapshot_for_test(receiver).await;
+    let (_, _, _, files_to_delete) = create_mooncake_snapshot_for_test(table, receiver).await;
     files_to_delete
 }
 
@@ -368,7 +371,7 @@ pub(crate) async fn perform_index_merge_for_test(
 ) -> Vec<String> {
     // Perform and block wait index merge.
     table.perform_index_merge(index_merge_payload);
-    let index_merge_result = MooncakeTable::sync_index_merge(receiver).await;
+    let index_merge_result = sync_index_merge(receiver).await;
 
     table.set_file_indices_merge_res(index_merge_result);
     assert!(table.create_snapshot(SnapshotOption {
@@ -377,7 +380,7 @@ pub(crate) async fn perform_index_merge_for_test(
         skip_file_indices_merge: false,
         skip_data_file_compaction: true,
     }));
-    let (_, _, _, evicted_files_to_delete) = MooncakeTable::sync_mooncake_snapshot(receiver).await;
+    let (_, _, _, evicted_files_to_delete) = sync_mooncake_snapshot(receiver).await;
     // Delete evicted object storage cache entries immediately to make sure later accesses all happen on persisted files.
     io_utils::delete_local_files(evicted_files_to_delete.clone())
         .await
@@ -396,8 +399,7 @@ pub(crate) async fn perform_data_compaction_for_test(
 ) -> Vec<String> {
     // Perform and block wait data compaction.
     table.perform_data_compaction(data_compaction_payload);
-    let data_compaction_result = MooncakeTable::sync_data_compaction(receiver).await;
-    let data_compaction_result = data_compaction_result.unwrap();
+    let data_compaction_result = sync_data_compaction(receiver).await;
 
     table.set_data_compaction_res(data_compaction_result);
     assert!(table.create_snapshot(SnapshotOption {
@@ -406,11 +408,271 @@ pub(crate) async fn perform_data_compaction_for_test(
         skip_file_indices_merge: true,
         skip_data_file_compaction: false,
     }));
-    let (_, _, _, evicted_files_to_delete) = MooncakeTable::sync_mooncake_snapshot(receiver).await;
+    let (_, _, _, evicted_files_to_delete) = sync_mooncake_snapshot(receiver).await;
     // Delete evicted object storage cache entries immediately to make sure later accesses all happen on persisted files.
     io_utils::delete_local_files(evicted_files_to_delete.clone())
         .await
         .unwrap();
 
     evicted_files_to_delete
+}
+
+/// ===================================
+/// Operation synchronization function
+/// ===================================
+///
+/// Test util function to block wait and get iceberg / file indices merge payload.
+async fn sync_mooncake_snapshot(
+    receiver: &mut Receiver<TableNotify>,
+) -> (
+    Option<IcebergSnapshotPayload>,
+    Option<FileIndiceMergePayload>,
+    Option<DataCompactionPayload>,
+    Vec<String>,
+) {
+    let notification = receiver.recv().await.unwrap();
+    if let TableNotify::MooncakeTableSnapshot {
+        iceberg_snapshot_payload,
+        file_indice_merge_payload,
+        data_compaction_payload,
+        evicted_data_files_to_delete,
+        ..
+    } = notification
+    {
+        (
+            iceberg_snapshot_payload,
+            file_indice_merge_payload,
+            data_compaction_payload,
+            evicted_data_files_to_delete,
+        )
+    } else {
+        panic!("Expected mooncake snapshot completion notification, but get others.");
+    }
+}
+async fn sync_iceberg_snapshot(receiver: &mut Receiver<TableNotify>) -> IcebergSnapshotResult {
+    let notification = receiver.recv().await.unwrap();
+    if let TableNotify::IcebergSnapshot {
+        iceberg_snapshot_result,
+    } = notification
+    {
+        iceberg_snapshot_result.unwrap()
+    } else {
+        panic!("Expected iceberg completion snapshot notification, but get mooncake one.");
+    }
+}
+async fn sync_index_merge(receiver: &mut Receiver<TableNotify>) -> FileIndiceMergeResult {
+    let notification = receiver.recv().await.unwrap();
+    if let TableNotify::IndexMerge { index_merge_result } = notification {
+        index_merge_result
+    } else {
+        panic!("Expected index merge completion notification, but get another one.");
+    }
+}
+async fn sync_data_compaction(receiver: &mut Receiver<TableNotify>) -> DataCompactionResult {
+    let notification = receiver.recv().await.unwrap();
+    if let TableNotify::DataCompaction {
+        data_compaction_result,
+    } = notification
+    {
+        data_compaction_result.unwrap()
+    } else {
+        panic!("Expected data compaction completion notification, but get mooncake one.");
+    }
+}
+
+/// ===================================
+/// Composite util functions
+/// ===================================
+// Test util function, which creates mooncake snapshot for testing.
+pub(crate) async fn create_mooncake_snapshot_for_test(
+    table: &mut MooncakeTable,
+    receiver: &mut Receiver<TableNotify>,
+) -> (
+    Option<IcebergSnapshotPayload>,
+    Option<FileIndiceMergePayload>,
+    Option<DataCompactionPayload>,
+    Vec<String>,
+) {
+    let mooncake_snapshot_created = table.create_snapshot(SnapshotOption {
+        force_create: true,
+        skip_iceberg_snapshot: false,
+        skip_data_file_compaction: false,
+        skip_file_indices_merge: false,
+    });
+    assert!(mooncake_snapshot_created);
+    sync_mooncake_snapshot(receiver).await
+}
+
+// Test util function, which updates mooncake table snapshot and create iceberg snapshot in a serial fashion.
+pub(crate) async fn create_mooncake_and_iceberg_snapshot_for_test(
+    table: &mut MooncakeTable,
+    receiver: &mut Receiver<TableNotify>,
+) {
+    // Create mooncake snapshot and block wait completion.
+    let (iceberg_snapshot_payload, _, _, evicted_data_files_to_delete) =
+        create_mooncake_snapshot_for_test(table, receiver).await;
+
+    // Delete evicted object storage cache entries immediately to make sure later accesses all happen on persisted files.
+    io_utils::delete_local_files(evicted_data_files_to_delete)
+        .await
+        .unwrap();
+
+    // Create iceberg snapshot if possible.
+    if let Some(iceberg_snapshot_payload) = iceberg_snapshot_payload {
+        table.persist_iceberg_snapshot(iceberg_snapshot_payload);
+        let iceberg_snapshot_result = sync_iceberg_snapshot(receiver).await;
+        table.set_iceberg_snapshot_res(iceberg_snapshot_result);
+    }
+}
+
+// Test util to block wait current mooncake snapshot completion, get the iceberg persistence payload, and perform a new mooncake snapshot and wait completion.
+async fn sync_mooncake_snapshot_and_create_new_by_iceberg_payload(
+    table: &mut MooncakeTable,
+    receiver: &mut Receiver<TableNotify>,
+) {
+    let (iceberg_snapshot_payload, _, _, evicted_data_files_to_delete) =
+        sync_mooncake_snapshot(receiver).await;
+    // Delete evicted object storage cache entries immediately to make sure later accesses all happen on persisted files.
+    io_utils::delete_local_files(evicted_data_files_to_delete)
+        .await
+        .unwrap();
+
+    let iceberg_snapshot_payload = iceberg_snapshot_payload.unwrap();
+    table.persist_iceberg_snapshot(iceberg_snapshot_payload);
+    let iceberg_snapshot_result = sync_iceberg_snapshot(receiver).await;
+    table.set_iceberg_snapshot_res(iceberg_snapshot_result);
+
+    // Create mooncake snapshot after buffering iceberg snapshot result, to make sure mooncake snapshot is at a consistent state.
+    assert!(table.create_snapshot(SnapshotOption {
+        force_create: true,
+        skip_iceberg_snapshot: true,
+        skip_file_indices_merge: true,
+        skip_data_file_compaction: true,
+    }));
+}
+
+// Test util function, which does the following things in serial fashion.
+// (1) updates mooncake table snapshot, (2) create iceberg snapshot, (3) trigger data compaction, (4) perform data compaction, (5) another mooncake and iceberg snapshot.
+//
+// # Arguments
+//
+// * injected_committed_deletion_rows: rows to delete and commit in between data compaction initiation and snapshot creation
+// * injected_uncommitted_deletion_rows: rows to delete but not commit in between data compaction initiation and snapshot creation
+#[cfg(test)]
+pub(crate) async fn create_mooncake_and_iceberg_snapshot_for_data_compaction_for_test(
+    table: &mut MooncakeTable,
+    receiver: &mut Receiver<TableNotify>,
+    injected_committed_deletion_rows: Vec<(MoonlinkRow, u64 /*lsn*/)>,
+    injected_uncommitted_deletion_rows: Vec<(MoonlinkRow, u64 /*lsn*/)>,
+) {
+    // Create mooncake snapshot.
+    let force_snapshot_option = SnapshotOption {
+        force_create: true,
+        skip_iceberg_snapshot: false,
+        skip_file_indices_merge: true,
+        skip_data_file_compaction: false,
+    };
+    assert!(table.create_snapshot(force_snapshot_option.clone()));
+
+    // Create iceberg snapshot.
+    let (iceberg_snapshot_payload, _, _, evicted_data_files_to_delete) =
+        sync_mooncake_snapshot(receiver).await;
+    // Delete evicted object storage cache entries immediately to make sure later accesses all happen on persisted files.
+    io_utils::delete_local_files(evicted_data_files_to_delete)
+        .await
+        .unwrap();
+
+    if let Some(iceberg_snapshot_payload) = iceberg_snapshot_payload {
+        table.persist_iceberg_snapshot(iceberg_snapshot_payload);
+        let iceberg_snapshot_result = sync_iceberg_snapshot(receiver).await;
+        table.set_iceberg_snapshot_res(iceberg_snapshot_result);
+    }
+
+    // Get data compaction payload.
+    assert!(table.create_snapshot(force_snapshot_option.clone()));
+    let (iceberg_snapshot_payload, _, data_compaction_payload, evicted_data_files_to_delete) =
+        sync_mooncake_snapshot(receiver).await;
+    // Delete evicted object storage cache entries immediately to make sure later accesses all happen on persisted files.
+    io_utils::delete_local_files(evicted_data_files_to_delete)
+        .await
+        .unwrap();
+
+    assert!(iceberg_snapshot_payload.is_none());
+    let data_compaction_payload = data_compaction_payload.unwrap();
+
+    // Perform and block wait data compaction.
+    table.perform_data_compaction(data_compaction_payload);
+    let data_compaction_result = sync_data_compaction(receiver).await;
+
+    // Before create snapshot for compaction results, perform another deletion operations.
+    for (cur_row, lsn) in injected_committed_deletion_rows {
+        table.delete(cur_row, lsn).await;
+        table.commit(/*lsn=*/ lsn);
+    }
+    for (cur_row, lsn) in injected_uncommitted_deletion_rows {
+        table.delete(cur_row, lsn).await;
+    }
+
+    // Set data compaction result and trigger another iceberg snapshot.
+    table.set_data_compaction_res(data_compaction_result);
+    assert!(table.create_snapshot(SnapshotOption {
+        force_create: true,
+        skip_iceberg_snapshot: false,
+        skip_file_indices_merge: true,
+        skip_data_file_compaction: false,
+    }));
+    sync_mooncake_snapshot_and_create_new_by_iceberg_payload(table, receiver).await;
+}
+
+// Test util function, which does the following things in serial fashion.
+// (1) updates mooncake table snapshot, (2) create iceberg snapshot, (3) trigger index merge, (4) perform index merge, (5) another mooncake and iceberg snapshot.
+pub(crate) async fn create_mooncake_and_iceberg_snapshot_for_index_merge_for_test(
+    table: &mut MooncakeTable,
+    receiver: &mut Receiver<TableNotify>,
+) {
+    // Create mooncake snapshot.
+    let force_snapshot_option = SnapshotOption {
+        force_create: true,
+        skip_iceberg_snapshot: false,
+        skip_file_indices_merge: false,
+        skip_data_file_compaction: false,
+    };
+    assert!(table.create_snapshot(force_snapshot_option.clone()));
+
+    // Create iceberg snapshot.
+    let (iceberg_snapshot_payload, _, _, evicted_data_files_to_delete) =
+        sync_mooncake_snapshot(receiver).await;
+    // Delete evicted object storage cache entries immediately to make sure later accesses all happen on persisted files.
+    io_utils::delete_local_files(evicted_data_files_to_delete)
+        .await
+        .unwrap();
+
+    if let Some(iceberg_snapshot_payload) = iceberg_snapshot_payload {
+        table.persist_iceberg_snapshot(iceberg_snapshot_payload);
+        let iceberg_snapshot_result = sync_iceberg_snapshot(receiver).await;
+        table.set_iceberg_snapshot_res(iceberg_snapshot_result);
+    }
+
+    // Perform index merge.
+    assert!(table.create_snapshot(force_snapshot_option.clone()));
+    let (iceberg_snapshot_payload, file_indice_merge_payload, _, evicted_data_files_to_delete) =
+        sync_mooncake_snapshot(receiver).await;
+    // Delete evicted object storage cache entries immediately to make sure later accesses all happen on persisted files.
+    io_utils::delete_local_files(evicted_data_files_to_delete)
+        .await
+        .unwrap();
+
+    assert!(iceberg_snapshot_payload.is_none());
+    let file_indice_merge_payload = file_indice_merge_payload.unwrap();
+
+    table.perform_index_merge(file_indice_merge_payload);
+    let index_merge_result = sync_index_merge(receiver).await;
+    table.set_file_indices_merge_res(index_merge_result);
+    assert!(table.create_snapshot(SnapshotOption {
+        force_create: true,
+        skip_iceberg_snapshot: false,
+        skip_file_indices_merge: false,
+        skip_data_file_compaction: false,
+    }));
+    sync_mooncake_snapshot_and_create_new_by_iceberg_payload(table, receiver).await;
 }
