+++ src/moonlink/src/storage/mooncake_table/tests.rs
@@ -1,6 +1,7 @@
 use super::test_utils::*;
 use super::*;
 use crate::storage::iceberg::table_manager::MockTableManager;
+use crate::storage::mooncake_table::state_test_utils::*;
 use iceberg::{Error as IcebergError, ErrorKind};
 use rstest::*;
 use rstest_reuse::{self, *};
@@ -173,9 +174,7 @@ async fn test_update_rows(#[case] identity: IdentityProp) -> Result<()> {
     table.append(row3.clone())?;
     table.commit(/*lsn=*/ 100);
     table.flush(/*lsn=*/ 100).await?;
-    table
-        .create_mooncake_and_iceberg_snapshot_for_test(&mut event_completion_rx)
-        .await?;
+    create_mooncake_and_iceberg_snapshot_for_test(&mut table, &mut event_completion_rx).await;
     {
         let mut table_snapshot = table.snapshot.write().await;
         let SnapshotReadOutput {
@@ -203,9 +202,7 @@ async fn test_update_rows(#[case] identity: IdentityProp) -> Result<()> {
     table.flush(/*lsn=*/ 300).await?;
 
     // Check update result.
-    table
-        .create_mooncake_and_iceberg_snapshot_for_test(&mut event_completion_rx)
-        .await?;
+    create_mooncake_and_iceberg_snapshot_for_test(&mut table, &mut event_completion_rx).await;
     {
         let mut table_snapshot = table.snapshot.write().await;
         let SnapshotReadOutput {
@@ -375,9 +372,7 @@ async fn test_duplicate_deletion() -> Result<()> {
     table.append(old_row.clone()).unwrap();
     table.commit(/*lsn=*/ 100);
     table.flush(/*lsn=*/ 100).await.unwrap();
-    table
-        .create_mooncake_and_iceberg_snapshot_for_test(&mut event_completion_rx)
-        .await?;
+    create_mooncake_and_iceberg_snapshot_for_test(&mut table, &mut event_completion_rx).await;
 
     // Update operation.
     let new_row = old_row.clone();
@@ -385,9 +380,7 @@ async fn test_duplicate_deletion() -> Result<()> {
     table.append(new_row.clone()).unwrap();
     table.commit(/*lsn=*/ 200);
     table.flush(/*lsn=*/ 200).await.unwrap();
-    table
-        .create_mooncake_and_iceberg_snapshot_for_test(&mut event_completion_rx)
-        .await?;
+    create_mooncake_and_iceberg_snapshot_for_test(&mut table, &mut event_completion_rx).await;
 
     {
         let mut table_snapshot = table.snapshot.write().await;
